Parallelize random Forest to PMML conversion
Just added the foreach loop with %dopar%, instead of lapply, so that it can parallelize when used with parallel backend runtimes of R. As an example, for my problem, Random forest with 200 trees was able to run in under 3 minutes, instead of the original 8-9 minutes. One dropback might be that you might not be able to see the message posted by the parallel process, depending on the different parallel backends.